Gestion de Informes y Abonados

+++ app/Http/Controllers/JugadorController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreJugadorRequest;
 use App\Http\Requests\UpdateJugadorRequest;
 use App\Models\Equipo;
 use App\Models\Jugador;
@@ -16,21 +15,26 @@ class JugadorController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $jugadores = Jugador::with('primera_posicion')->get();
+        $equipoId = $request->input('equipo');
+
+        $jugadores = Jugador::with('primera_posicion')
+            ->when($equipoId, function ($query, $equipoId) {
+                $query->where('equipo_id', $equipoId);
+            })
+            ->get();
 
         return Inertia::render('Jugadores/Index', [
-            'jugadores' => $jugadores
+            'jugadores' => $jugadores,
         ]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
-
         $equipos = Equipo::all();
         $posiciones = Posicion::all();
         $representantes = Representante::all();
@@ -48,6 +52,7 @@ class JugadorController extends Controller
     public function store(Request $request)
     {
 
+        $equipoId = $request->input('equipo');
         $validated = $request->validate([
             'apodo' => 'required|string|max:255',
             'nombre' => 'required|string|max:255',
@@ -81,7 +86,7 @@ class JugadorController extends Controller
         }
         Jugador::create($validated);
 
-        return redirect()->route('equipos.index');
+        return redirect()->route('jugadores.index', $equipoId)->with('success', 'Jugador creado correctamente.');
     }
 
     /**
@@ -97,6 +102,7 @@ class JugadorController extends Controller
      */
     public function edit(Jugador $jugador)
     {
+
         $equipos = Equipo::all();
         $posiciones = Posicion::all();
         $representantes = Representante::all();
@@ -112,7 +118,7 @@ class JugadorController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateJugadorRequest $request, Jugador $jugador)
+    public function update(Request $request, Jugador $jugador)
     {
         $validated = $request->validate([
             'apodo' => 'required|string|max:255',
@@ -142,7 +148,7 @@ class JugadorController extends Controller
 
         $jugador->update($validated);
 
-        return redirect()->route('jugadores.index');
+        return redirect()->route('jugadores.index') ->with('success', 'Jugador modificado correctamente.');
     }
 
     /**
@@ -150,8 +156,13 @@ class JugadorController extends Controller
      */
     public function destroy(Jugador $jugador)
     {
+
+        if ($jugador->informes()->exists()) {
+            return redirect()->route('jugadores.index')->with('error', 'No se puede eliminar el jugador porque tiene informes asociados.');
+        }
+
         $jugador->delete();
 
-        return redirect()->route('jugadores.index');
+        return redirect()->route('jugadores.index')->with('success', 'Jugador eliminado correctamente.');
     }
 }
